Add files via upload
Link Dataset: https://www.kaggle.com/datasets/edumagalhaes/quality-prediction-in-a-mining-process

Project description: Forecasting quality in the mining process
The purpose of the presentation
This work is devoted to the creation of a machine learning model for predicting quality at the stages of the production process in the mining industry. Using modern data analysis and machine learning tools, the project makes it possible to increase the accuracy of quality prediction, which in the long term contributes to improving the efficiency of the entire process.